+++ simplesockets/__init__.py
@@ -1,0 +1,3 @@
+from .simple_sockets import TCPServer, TCPClient
+from .securesockets.SelfmadeCipher_Socket import SecureClient as PySecureClient, SecureServer as PySecureServer
+from .securesockets.PyCryptodome_Socket import SecureClient, SecureServer