+++ simplesockets/_support_files/RSA.py
@@ -2,6 +2,7 @@ import secrets
 import random
 import os
 from typing import Union
+import json
 
 from simplesockets._support_files import error
 
@@ -214,13 +215,82 @@ def import_public_bytes(key_info: bytes):
     e = e.strip(b'e=')
     return RSA_Public_Key(int(n), int(e))
 
+def get_prime_list():
+    with open('primes.json') as primes_file:
+        return json.load(primes_file)
 
-def get_private_key(key_length: int = 4096) -> RSA_Private_Key:
+def get_single_key(primes_list, length: int = 2048):
+    """
+    generates a prime number
+
+    Args:
+        primes_list: list of primes
+        length: the size of the prime in bits
+
+    Returns:
+        returns an prime int
+
+    """
+    def random_bit_number_odd(size: int = 2048):
+        l = secrets.randbits(size)
+        while pow(l, 1, 2) == 0:
+            l = secrets.randbits(size)
+        return l
+
+    def check_first_primes(n: int):
+        for prime_ in primes_list:
+            if pow(n, 1, prime_) == 0:
+                return False
+        return True
+
+    def fermit_test(n: int, k: int = 10):
+        for _ in range(k):
+            a = random.randint(2, n - 2)
+            pow_ = pow(a, n - 1, n)
+            if pow_ != 1:
+                return False
+        return True
+
+    def Miller_Rabin_test(n: int, k: int):
+        r, s = 0, n - 1
+        while pow(s, 1, 2) == 0:
+            r += 1
+            s //= 2
+        for _ in range(k):
+            a = random.randrange(2, n - 1)
+            x = pow(a, s, n)
+            if x == 1 or x == n - 1:
+                continue
+            for _ in range(r - 1):
+                x = pow(x, 2, n)
+                if x == n - 1:
+                    break
+            else:
+                return False
+        return True
+
+    prime = False
+
+    while not prime:
+        number = random_bit_number_odd(length)
+        if check_first_primes(number) is False:
+            continue
+        if fermit_test(number, 5) is False:
+            continue
+        if Miller_Rabin_test(number, 5) is False:
+            continue
+        prime = True
+
+    return number
+
+def get_private_key(key_length: int = 4096, p: int = None, q: int = None) -> RSA_Private_Key:
     """
     Creates and Returns a RSA Public Key
 
     Args:
         key_length: bit length of the key, it should be even
+        p: when given, p will be used as p in calculating the RSA Key
+        q: when given, q will be used as q in calculating the RSA Key
 
     Returns:
         returns a RSA Private Key
@@ -230,86 +300,11 @@ def get_private_key(key_length: int = 4096) -> RSA_Private_Key:
         RSACalcKeyError: If calculating d of the RSA Key resulted in an invalid d value
 
     """
-    primes_list = {3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53, 59, 61, 67, 71, 73, 79, 83, 89, 97, 101,
-                   103, 107, 109, 113, 127, 131, 137, 139, 149, 151, 157, 163, 167, 173, 179, 181, 191, 193, 197, 199,
-                   211, 223, 227, 229, 233, 239, 241, 251, 257, 263, 269, 271, 277, 281, 283, 293, 307, 311, 313, 317,
-                   331, 337, 347, 349, 353, 359, 367, 373, 379, 383, 389, 397, 401, 409, 419, 421, 431, 433, 439, 443,
-                   449, 457, 461, 463, 467, 479, 487, 491, 499, 503, 509, 521, 523, 541, 547, 557, 563, 569, 571, 577,
-                   587, 593, 599, 601, 607, 613, 617, 619, 631, 641, 643, 647, 653, 659, 661, 673, 677, 683, 691, 701,
-                   709, 719, 727, 733, 739, 743, 751, 757, 761, 769, 773, 787, 797, 809, 811, 821, 823, 827, 829, 839,
-                   853, 857, 859, 863, 877, 881, 883, 887, 907, 911, 919, 929, 937, 941, 947, 953, 967, 971, 977, 983,
-                   991, 997, 1009, 1013, 1019, 1021, 1031, 1033, 1039, 1049, 1051, 1061, 1063, 1069, 1087, 1091, 1093,
-                   1097, 1103, 1109, 1117, 1123, 1129, 1151, 1153, 1163, 1171, 1181, 1187, 1193, 1201, 1213, 1217, 1223,
-                   1229, 1231, 1237, 1249, 1259, 1277, 1279, 1283, 1289, 1291, 1297, 1301, 1303, 1307, 1319, 1321, 1327,
-                   1361, 1367, 1373, 1381, 1399, 1409, 1423, 1427, 1429, 1433, 1439, 1447, 1451, 1453, 1459, 1471, 1481,
-                   1483, 1487, 1489, 1493, 1499, 1511, 1523, 1531, 1543, 1549, 1553, 1559, 1567, 1571, 1579, 1583, 1597,
-                   1601, 1607, 1609, 1613, 1619, 1621, 1627, 1637, 1657, 1663, 1667, 1669, 1693, 1697, 1699, 1709, 1721,
-                   1723, 1733, 1741, 1747, 1753, 1759, 1777, 1783, 1787, 1789, 1801, 1811, 1823, 1831, 1847, 1861, 1867,
-                   1871, 1873, 1877, 1879, 1889, 1901, 1907, 1913, 1931, 1933, 1949, 1951, 1973, 1979, 1987, 1993, 1997,
-                   1999, 2003, 2011, 2017, 2027, 2029, 2039, 2053, 2063, 2069, 2081, 2083, 2087, 2089, 2099, 2111, 2113,
-                   2129, 2131, 2137, 2141, 2143, 2153, 2161, 2179, 2203, 2207, 2213, 2221, 2237, 2239, 2243, 2251, 2267,
-                   2269, 2273, 2281, 2287, 2293, 2297, 2309, 2311, 2333, 2339, 2341, 2347, 2351, 2357, 2371, 2377, 2381,
-                   2383, 2389, 2393, 2399, 2411, 2417, 2423, 2437, 2441, 2447, 2459, 2467, 2473, 2477, 2503, 2521, 2531,
-                   2539, 2543, 2549, 2551, 2557, 2579, 2591, 2593, 2609, 2617, 2621, 2633, 2647, 2657, 2659, 2663, 2671,
-                   2677, 2683, 2687, 2689, 2693, 2699, 2707, 2711, 2713, 2719, 2729, 2731, 2741, 2749, 2753, 2767, 2777,
-                   2789, 2791, 2797, 2801, 2803, 2819, 2833, 2837, 2843, 2851, 2857, 2861, 2879, 2887, 2897, 2903, 2909,
-                   2917, 2927, 2939, 2953, 2957, 2963, 2969, 2971, 2999, 3001, 3011, 3019, 3023, 3037, 3041, 3049, 3061,
-                   3067, 3079, 3083, 3089, 3109, 3119, 3121, 3137, 3163, 3167, 3169, 3181, 3187, 3191, 3203, 3209, 3217,
-                   3221, 3229, 3251, 3253, 3257, 3259, 3271, 3299, 3301, 3307, 3313, 3319, 3323, 3329, 3331, 3343, 3347,
-                   3359, 3361, 3371, 3373, 3389, 3391, 3407, 3413, 3433, 3449, 3457, 3461, 3463, 3467, 3469, 3491, 3499,
-                   3511, 3517, 3527, 3529, 3533, 3539, 3541, 3547, 3557, 3559, 3571}
+
+    primes_list = get_prime_list()
 
     if key_length % 2 != 0:
         raise ValueError("key length must be even")
-
-    def get_single_key(length: int = 2048):
-        def random_bit_number_odd(size: int = 4096):
-            l = secrets.randbits(size)
-            while l % 2 == 0:
-                l = secrets.randbits(size)
-            return l
-
-        def check_first_primes(n: int):
-            return n not in primes_list
-
-        def fermit_test(n: int, k: int = 10):
-            for _ in range(k):
-                a = random.randint(2, n - 2)
-                pow_ = pow(a, n - 1, n)
-                if pow_ != 1:
-                    return False
-            return True
-
-        def Miller_Rabin_test(n: int, k: int):
-            r, s = 0, n - 1
-            while s % 2 == 0:
-                r += 1
-                s //= 2
-            for _ in range(k):
-                a = random.randrange(2, n - 1)
-                x = pow(a, s, n)
-                if x == 1 or x == n - 1:
-                    continue
-                for _ in range(r - 1):
-                    x = pow(x, 2, n)
-                    if x == n - 1:
-                        break
-                else:
-                    return False
-            return True
-
-        prime = False
-        while not prime:
-            number = random_bit_number_odd(length)
-            if check_first_primes(number) is False:
-                continue
-            if fermit_test(number, 5) is False:
-                continue
-            if Miller_Rabin_test(number, 5) is False:
-                continue
-            prime = True
-
-        return number
 
     def get_e(phi_: int):
         e = 65537
@@ -335,8 +330,14 @@ def get_private_key(key_length: int = 4096) -> RSA_Private_Key:
             if c1 == 1:
                 return c2
 
-    p = get_single_key(key_length // 2)
-    q = get_single_key(key_length // 2)
+    if isinstance(p, int) is False:
+        p = get_single_key(primes_list, key_length // 2)
+    if isinstance(q, int) is False:
+        q = get_single_key(primes_list, key_length // 2)
+
+    if p == q:
+        raise ValueError("p and q can't have the same Value")
+
     n = p * q
     phi = (p - 1) * (q - 1)
     e = get_e(phi)
@@ -388,3 +389,12 @@ def import_key(key: bytes) -> Union[RSA_Public_Key, RSA_Private_Key]:
         return RSA_Public_Key(int(info[0]), int(info[1]))
     else:
         raise error.RSAImportKeyError("Couldn't identify key")
+
+if __name__ == "__main__":
+    import time
+    start_time = time.time()
+    key = get_private_key(1024)
+    end_time = time.time() - start_time
+
+    print(key)
+    print(f"time taken: {end_time} seconds, {end_time*1000} milliseconds, {end_time/60} minutes")