+++ simplesockets/_support_files/error.py
@@ -21,6 +21,18 @@ class VPadError(VeginerError): pass
 class GenerationTypeError(VeginerError): pass
 
 
+class TranspositionError(error):
+    pass
+
+class NotEncryptedError(TranspositionError): pass
+
+class InvalidKeyError(TranspositionError): pass
+
+class RepeatingValueError(InvalidKeyError): pass
+
+class NoEffectError(InvalidKeyError): pass
+
+
 class SocketError(error):
     pass